// DXC_Frontend/src/api/models/bookingOrderDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface BookingOrderDto {
  publicId?: string
  /** @nullable */
  bookingCode?: string | null
  /** @nullable */
  customerName?: string | null
  /** @nullable */
  phoneNumber?: string | null
  /** @nullable */
  email?: string | null
  /** @nullable */
  note?: string | null
  /** @nullable */
  tourId?: string | null
  /** @nullable */
  ticketId?: string | null
  /** @nullable */
  serviceType?: string | null
  /** @nullable */
  serviceName?: string | null
  quantity?: number
  adultQuantity?: number
  childQuantity?: number
  /** @nullable */
  departureDate?: string | null
  /** @nullable */
  departureTime?: string | null
  totalAmount?: number
  /** @nullable */
  status?: string | null
  /** @nullable */
  paymentStatus?: string | null
  createdAt?: string
  updatedAt?: string
}